added two variables for redir and subshell

+++ src/executor/execute_tree.c
@@ -54,12 +54,30 @@ int	exec_ctrl(t_ast_node *node)
 */
 int	exec_block(t_ast_node *node)
 {
-	int	(*builtin)(t_ast_node *node);
+	int	status_;
 
-	builtin = is_builtin(node->args[0]);
-	if (builtin)
-		return (set_exit_status(builtin(node)), get_exit_status());
-	return (1); //temprory
+	status_= 0;
+	signal(SIGINT, interrput_slience);
+	signal(SIGQUIT, interrput_slience);
+	if (fork() == -1)
+		return (error("fork() failed", NULL), EXIT_FAILURE);
+	else
+	{
+		signal(SIGINT, SIG_DFL);
+		signal(SIGQUIT, SIG_DFL);
+		status_= executor_status(node->subshell_cmd);
+		set_exit_status(status_);
+		free_all_memory();
+		exit(status_);
+	}
+	wait(&status_);
+	if (WIFEXITED(status_))
+		status_ = WEXITSTATUS(status_);
+	else
+		status_ = EXIT_FAILURE;
+	set_exit_status(status_);
+	signals_init();
+	return (status_);
 }
 
 /*
@@ -102,6 +120,7 @@ int	exec_redir(t_ast_node *node)
 	t_redirection	*redir;
 	int				status_;
 	int				fd;
+	int				dup_fd;
 	int				flags;
 
 	if (!node || !node->redirections)
@@ -112,13 +131,20 @@ int	exec_redir(t_ast_node *node)
 		flags = get_redirection_flags(node->type);
 		fd = open(redir->file, flags, 0644);
 		if (fd == -1)
-		{
-			error(redir->file, NULL);
-			return (set_exit_status(1), 1);
-		}
-		status_ = executor_status(node);
+			return (error(redir->file, NULL), set_exit_status(1), 1);
+		dup_fd = dup(get_redirection_fd(redir->type));
+		if (dup_fd == -1)
+			return (error("dup failed", NULL), set_exit_status(1), 1);
+		if (dup2(fd, get_redirection_fd(redir->type)) == -1)
+			return (error("dup2 failed", NULL), set_exit_status(1), 1);
+		close(fd);
+		status_ = executor_status(node->redir_cmd);
+		if (dup2(dup_fd, get_redirection_fd(redir->type)) == -1)
+			return (error("dup failed", NULL), set_exit_status(1), 1);
+		close(dup_fd);
 		if (redir->type == TOKEN_HEREDOC)
 			unlink(redir->file);
+		redir = redir->next;
 	}
 	return (status_);
 }
